TapMania lives!
Removed all tracking and ads
Disabled (for the time being) uploading new songs via HTTP
Removed support for OGG format (it is not hardware accelerated)
Removed (for the time being) iCade support
Updated some code to not use deprecated functionality
Fixed compiler warnings
Updated README

+++ GameObjects/TMTrack.h
@@ -40,7 +40,7 @@ typedef vector<TMNotePtr> TMNoteList;
 
 - (BOOL) hasNoteAtRow:(int)noteRow;
 
-- (int) getNotesCount;
+- (unsigned long) getNotesCount;
 - (int) getHoldsCount;
 - (int) getTapAndHoldNotesCount;
